backend/ecotruck/vehicle/serializers.py
from rest_framework import serializers
from core.models import Vehicle


class VehicleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Vehicle
        fields = ["id", "name", "brand", "model"]
