backend/pyproject.toml
[project]
name = "ecotruck-app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "django>=4.2",
  "djangorestframework>=3.14",
  "django-filter>=25.2"
]
